Move "attach source" menu into own comp
