+++ application/client/src/app/ui/elements/menu.attachsource/styles.less
@@ -1,0 +1,17 @@
+@import '../../styles/variables.less';
+
+app-attach-new-source-menu {
+    display: inline-block;
+}
+p.attach-menu-warning {
+    padding: 12px 24px;
+    color: @scheme-color-1;
+}
+div.attach-new-source-menu {
+    & .mat-icon {
+        width: 16px;
+        height: 16px;
+        font-size: 16px;
+        color: @scheme-color-1;
+    }
+}
