+++ application/client/src/app/ui/views/workspace/title/template.html
@@ -1,7 +1,5 @@
 <span class="t-normal title">{{state.title}}</span>
-<!-- <span
-    title="Add new source" 
-    class="with-background small-icon-button codicon codicon-diff-added"></span> -->
+<app-attach-new-source-menu [session]="session"></app-attach-new-source-menu>
 <span
     title="Sending data into stream" 
     [attr.class]="'with-background small-icon-button codicon codicon-terminal' + (state.flags.sde ? ' enabled' : ' disabled') + (state.states.sde ? ' active' : ' inactive')" (click)="state.toggleSde()"></span>
